requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
discord==2.3.2
discord.py==2.4.0
feedparser==6.0.11
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
protobuf==4.21.12
python-dotenv==1.0.1
pywatchman==1.4.1
sgmllib3k==1.0.0
six==1.16.0
yarl==1.9.4
